+++ android/src/main/kotlin/net/mullvad/mullvadvpn/model/TunnelOptions.kt
@@ -1,0 +1,5 @@
+package net.mullvad.mullvadvpn.model
+
+import net.mullvad.talpid.net.wireguard.TunnelOptions as WireguardTunnelOptions
+
+data class TunnelOptions(val wireguard: WireguardTunnelOptions)